Write a Kafka producer, publishing to "my_orders" topic
This is the first lab exercise.
Of special note is the shutdown hook for properly cleaning up
the producer resources (flush and close).  The producer is
thread-safe, so it is OK to do this from a new thread created
by the shutdown hook.

Run in your IDE, or run in Maven using:
```
mvn exec:java -Dexec.mainClass="com.xyzcorp.MyProducer"
```

Use the broker (cp-server) container shell to consume messages
from the topic (my_orders):
```
kafka-console-consumer \
  --bootstrap-server localhost:9092 \
  --topic my_orders --from-beginning \
  --key-deserializer org.apache.kafka.common.serialization.StringDeserializer \
  --value-deserializer org.apache.kafka.common.serialization.IntegerDeserializer \
  --property print.key=true \
  --property key.separator=,
```